% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.linear.R
\name{model.linear}
\alias{model.linear}
\title{Run Linear Models}
\usage{
model.linear(n, phenotype, odata)
}
\arguments{
\item{n}{Number of iterations to run}

\item{phenotype}{A variable representing phenotype- which is used a reponse variable for linear regression}

\item{odata}{Count data with rows as samples, and columns as features/OTUs/ASVs}
}
\value{
Features from linear regression
}
\description{
This function takes in phyloseq object, association matix, p value matrix
and create a combined OTU-OTU and OTU-Phenotype network. User can select model to
define OTU-Phontype assocaition.
}
