% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noderoles.R
\name{node.role}
\alias{node.role}
\title{Run Node Roles using rnetcarto}
\usage{
node.role(graph_object)
}
\arguments{
\item{graph_object}{A graph object from igraph.}
}
\value{
A plot with nodes classified into four groups based on within-module degree and among-module connectivity
}
\description{
This function takes in graph_object from igraph, and run rnetcarto function to assign roles based SA algorithm.
}
