% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.mothur.shared.R
\name{read.mothur.shared}
\alias{read.mothur.shared}
\title{Reads .shared file from mothur}
\usage{
read.mothur.shared(shared.file)
}
\arguments{
\item{shared.file}{A .shared file}
}
\value{
A table with rows as features(OTUs) and columns as samples
}
\description{
Reads in .shared file, a count matrix from mothur.
}
