% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removerare.R
\name{remove_rare}
\alias{remove_rare}
\title{Remove rare observation based on prevalence percentage and/or minimum count}
\usage{
remove_rare(x, minOTUcount = 10, prevalence_percent = 5)
}
\arguments{
\item{x}{Vector of observations for a feature or a column of a feature from a feature matrix.}
}
\value{
Boolean. If a feature is within the defined criteria, return True, else False
}
\description{
This function removes rare feature based on revalence percentage and/or minimum count
}
