% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxacolor.R
\name{taxacolor}
\alias{taxacolor}
\title{Function to assign color to the phyloseq object}
\usage{
taxacolor(phyobj, coloredby = "x")
}
\arguments{
\item{phyobj}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{coloredby}{Select a taxonomic rank for coloring nodes in the network diagram. Possible option: "Kingdom","Phylum","Class","Order","Family","Genus","Species"}
}
\value{
phyobj2 An updated phyloseq object with color information
}
\description{
This function takes in phyloseq object, association matix, p value matrix
and create a combined OTU-OTU and OTU-Phenotype network. User can select model to
define OTU-Phontype assocaition.
}
