% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso}
\alias{lasso}
\title{Run Lasso Models}
\usage{
lasso(Phenotype, odata)
}
\arguments{
\item{Phenotype}{A phenotype variable to use for lasso regression}
}
\value{
odata A count data for feature, with rows as sample, and columns as features/ASVs/OTUs
}
\description{
This function takes in phenotype information and count data to run lasso regression. For hyper paramter optimization, 5 fold cv method is applied.
}
