% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhONA.R
\name{PhONA}
\alias{PhONA}
\title{Run PhONA}
\usage{
PhONA(
  physeqobj = physeq,
  cordata = sparcc.cor,
  pdata = sparcc.pval,
  model = "lm",
  OTU_OTU_pvalue = 0.001,
  OTU_OTU_rvalue = 0.6,
  OTU_Phenotype_pvalue = 0.6,
  definePhenotype = "Marketable",
  defineTreatment = "Maxifort",
  coloredby = "Phylum",
  PhenoNodecolor = "yellow",
  PhenoNodesize = 20,
  PhenoNodelabel = "Yield",
  nodesize = 10,
  Pheno2OTUedgecolor = "black",
  netlayout = layout.fruchterman.reingold
)
}
\arguments{
\item{physeqobj}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{cordata}{A pairwise square matrix defining OTU-OTU association}

\item{pdata}{A pairwise square matrix defining p-value for OTU-OTU association}

\item{model}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{OTU_OTU_pvalue}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{OTU_OTU_rvalue}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{OTU_Phenotype_pvalue}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{definePhenotype}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{defineTreatment}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{coloredby}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{PhenoNodecolor}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{PhenoNodesize}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{nodesize}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{Pheno2OTUedgecolor}{A phyloseq object which combined OTU count, taxonomy and metadata}

\item{netlayout}{A phyloseq object which combined OTU count, taxonomy and metadata}
}
\value{
A matrix of the infile
}
\description{
This function takes in phyloseq object, association matix, p value matrix
and create a combined OTU-OTU and OTU-Phenotype network. User can select model to
define OTU-Phontype assocaition.
}
