% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.mothur.taxonomy.R
\name{read.mothur.taxonomy}
\alias{read.mothur.taxonomy}
\title{Reads taxonomy file from mothur}
\usage{
read.mothur.taxonomy(tax.file)
}
\arguments{
\item{tax.file}{A taxonomy file obtained from mothur}
}
\value{
A table of formatted taxonomy file
}
\description{
Reads taxonomy file from mothur and remove all the symbols such as quotation, comma, semicolon etc
}
